wapow-collector/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0
clickhouse-connect>=0.8.0
pydantic>=2.0
python-dotenv>=1.0.0
geoip2>=4.8.0
